complete file and group function, update ui design

+++ src/utils/UserInfo.h
@@ -6,7 +6,7 @@
 
 class UserInfo
 {
-public:
+   public:
     // 删除拷贝构造函数和赋值操作符，确保单例
     UserInfo(const UserInfo&) = delete;
     UserInfo& operator=(const UserInfo&) = delete;
@@ -15,20 +15,20 @@ public:
     static UserInfo& instance();
 
     // 用户信息设置
-    void setUserId(int id) { m_userId = id; }
+    void setUserId(long id) { m_userId = id; }
     void setUsername(const QString& username) { m_username = username; }
     void setNickname(const QString& nickname) { m_nickname = nickname; }
     void setToken(const QString& token) { m_token = token; }
 
     // 用户信息获取
-    int userId() const { return m_userId; }
+    long userId() const { return m_userId; }
     QString username() const { return m_username; }
     QString nickname() const { return m_nickname; }
     QString token() const { return m_token; }
 
     // 检查是否已登录
     bool isLoggedIn() const { return !m_token.isEmpty(); }
-    
+
     // 清空用户信息（登出时使用）
     void clear()
     {
@@ -38,14 +38,14 @@ public:
         m_token.clear();
     }
 
-private:
-    UserInfo() = default;  // 私有构造函数
-    ~UserInfo() = default; // 私有析构函数
+   private:
+    UserInfo() = default;   // 私有构造函数
+    ~UserInfo() = default;  // 私有析构函数
 
-    int m_userId = -1;
+    long m_userId = -1;
     QString m_username;
     QString m_nickname;
     QString m_token;
 };
 
-#endif // USERINFO_H
+#endif  // USERINFO_H
